requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0
pydantic>=2.5
pydantic-settings>=2.4
python-jose>=3.3.0
bcrypt>=4.0
python-multipart>=0.0.6
pypdf>=3.0
pdfplumber>=0.9.1
reportlab>=4.0
requests>=2.30
numpy>=1.27
streamlit>=1.26
pytest>=8.0
httpx>=0.27
